OCPBUGS-31570: Make sure thin provisioning DMs are cleaned correctly
When a thin provisioning DM is listed, only the `_tdata` and `_tmeta` parts of that provisioning are listed as associated with a given disk.
This results in the disk cleanup trying to delete the `_tdata` and `_tmeta` parts before deleting the base part.

This PR changes that by deriving the base part so that it may be deleted prior to removal of `_tmeta` and `_tdata` parts of the provisioning.
This also ensures that the base DMs will always be listed first.

Already tested this by building against master and linking in go.mod of assisted-installer, then using this for a cluster installation in which we manually create a thin provisioning.
The provisioning is correctly cleaned and the installation completes, where it failed to do so before.